requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
uvicorn
fastapi
langserve
sse_starlette